feat(ui): consolidate NAS Devices and Device Registry into single tabbed page
- Create NetworkDevices.jsx as unified page with tabs
- Extract NasTab and DeviceRegistryTab components from existing pages
- Add info box explaining Device Registry purpose (RADIUS-proxy MAC lookup)
- Add /devices route with redirects from legacy /nas and /device-registry
- Merge navigation into single 'Network Devices' entry in Layout
- Re-export legacy page components for backward compatibility
- Category name visible in Device Registry table
